minor - rewrite grouping with stream's groupingby

+++ src/main/java/hipravin/jarvis/engine/model/AuthorResult.java
@@ -2,7 +2,7 @@ package hipravin.jarvis.engine.model;
 
 public record AuthorResult(
         String author,
-        int count
+        long count
 ) {
     public static final String TOTAL = "Total";
     public static final String OTHERS = "others";
